simulators/eth2/dencun: code cleanup
